requirements.txt
alembic==1.13.1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.2
msgspec==0.18.6
packaging==24.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
